a limit of 255 bar numbers max might probably be too small
Long musics might exceeds this limit. Let's increase to 65535 maximum
Note: limit is 65535 instead of 65536 since bar numbers start at 1.
Hence the value 0 can't be taken. Complicating the code by puting +1
everywhere isn't worth it.

+++ src/notes_file_extractor.hh
@@ -42,7 +42,7 @@ struct note_t
 {
     uint64_t start_time;
     uint64_t stop_time;
-    uint8_t bar_number;
+    uint16_t bar_number;
     pitch_t pitch;
     bool is_played;
     std::string id;
